Custom Button Property Changes
Implemented a delegate that is passed into the "CustomGroupBox" class that acts like a callback function for whenever the user changes a values in the custom button property panels. The delegate points to the "ApplyChanges" function located inside of the "CustomButton" class. Still need to figure out how to apply the changes correctly but for now it seems to be working fine.